// src/shared/open-api/models/requestEmailOtpResponseDto.ts
 
/**
 * Generated by orval v1.0 🍺
 * Do not edit manually.
 * Micro App API
 */

export interface RequestEmailOtpResponseDto {
  /** Status message */
  message: string;
  /** Seconds until code expires */
  expiresIn: number;
  /** True if no profile exists for this email */
  isNewUser: boolean;
}